Use OpenSSL as primitives. Also added some javadoc.
