+++ src/main/java/dk/alexandra/fresco/suite/tinytables/util/ot/OTSender.java
@@ -6,6 +6,11 @@ import dk.alexandra.fresco.suite.tinytables.util.ot.datatypes.OTInput;
 
 public interface OTSender {
 
+	/**
+	 * Perform oblivious transfer as the sender with the given inputs.
+	 * 
+	 * @param inputs
+	 */
 	public void send(List<OTInput> inputs);
-	
+
 }
