+++ src/main/java/dk/alexandra/fresco/suite/tinytables/util/ot/java/JavaOTReceiver.java
@@ -1,7 +1,6 @@
 package dk.alexandra.fresco.suite.tinytables.util.ot.java;
 
 import java.io.IOException;
-import java.io.Serializable;
 import java.security.SecureRandom;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -11,7 +10,6 @@ import dk.alexandra.fresco.framework.network.Network;
 import dk.alexandra.fresco.suite.tinytables.util.Encoding;
 import dk.alexandra.fresco.suite.tinytables.util.ot.OTReceiver;
 import dk.alexandra.fresco.suite.tinytables.util.ot.datatypes.OTSigma;
-import edu.biu.scapi.comm.Channel;
 import edu.biu.scapi.exceptions.FactoriesException;
 import edu.biu.scapi.exceptions.SecurityLevelException;
 import edu.biu.scapi.interactiveMidProtocols.ot.otBatch.OTBatchOnByteArrayROutput;
@@ -19,27 +17,38 @@ import edu.biu.scapi.interactiveMidProtocols.ot.otBatch.OTBatchRBasicInput;
 import edu.biu.scapi.interactiveMidProtocols.ot.otBatch.semiHonest.OTSemiHonestDDHBatchOnByteArrayReceiver;
 import edu.biu.scapi.tools.Factories.KdfFactory;
 
+/**
+ * This OTReciever is a wrapper around SCAPI's
+ * {@link OTSemiHonestDDHBatchOnByteArrayReceiver} based on an elliptic curve
+ * over a finite field, namely the K-163 curve.
+ * 
+ * @author Jonas Lindstrøm (jonas.lindstrom@alexandra.dk)
+ *
+ */
 public class JavaOTReceiver implements OTReceiver {
 
 	private Network network;
 	private int myId;
 	private SecureRandom random;
 
+	/*
+	 * We keep a singleton of the actual receiver.
+	 */
 	private static OTSemiHonestDDHBatchOnByteArrayReceiver receiver;
 	private OTSemiHonestDDHBatchOnByteArrayReceiver getInstance(SecureRandom random) {
 		if (receiver == null) {
 			try {
 				receiver = new OTSemiHonestDDHBatchOnByteArrayReceiver(
-						new edu.biu.scapi.primitives.dlog.bc.BcDlogECF2m(), KdfFactory.getInstance()
+						//new edu.biu.scapi.primitives.dlog.bc.BcDlogECF2m(),
+						new edu.biu.scapi.primitives.dlog.openSSL.OpenSSLDlogECF2m(), 
+						KdfFactory.getInstance()
 						.getObject("HKDF(HMac(SHA-256))"), random);
-			} catch (SecurityLevelException | IOException | FactoriesException e) {
+			} catch (SecurityLevelException | FactoriesException | IOException e) {
 				e.printStackTrace();
 			}
 		} 
 		return receiver;
 	}
-	
-	
 	
 	public JavaOTReceiver(Network network, int myId, SecureRandom random) {
 		this.network = network;
@@ -59,29 +68,8 @@ public class JavaOTReceiver implements OTReceiver {
 		OTBatchRBasicInput input = new OTBatchRBasicInput(s);
 		
 		try {
-			OTBatchOnByteArrayROutput output = (OTBatchOnByteArrayROutput) receiver.transfer(new Channel() {
-
-				@Override
-				public void close() {
-					
-				}
-
-				@Override
-				public boolean isClosed() {
-					return false;
-				}
-
-				@Override
-				public Serializable receive() throws ClassNotFoundException, IOException {
-					return network.receive("0", otherPlayerId());
-				}
-
-				@Override
-				public void send(Serializable arg0) throws IOException {
-					network.send("0", otherPlayerId(), arg0);
-				}
-				
-			}, input);
+			OTBatchOnByteArrayROutput output = (OTBatchOnByteArrayROutput) receiver.transfer(
+					new NetworkWrapper(network, myId), input);
 			
 			List<boolean[]> results = new ArrayList<boolean[]>();
 			for (int i = 0; i < sigmas.size(); i++) {
@@ -94,10 +82,6 @@ public class JavaOTReceiver implements OTReceiver {
 		}
 		
 		return null;
-	}
-	
-	private int otherPlayerId() {
-		return myId == 1 ? 2 : 1;
 	}
 
 }
