+++ src/main/java/dk/alexandra/fresco/suite/tinytables/util/ot/java/JavaOTSender.java
@@ -1,7 +1,6 @@
 package dk.alexandra.fresco.suite.tinytables.util.ot.java;
 
 import java.io.IOException;
-import java.io.Serializable;
 import java.security.SecureRandom;
 import java.util.ArrayList;
 import java.util.List;
@@ -10,30 +9,43 @@ import dk.alexandra.fresco.framework.network.Network;
 import dk.alexandra.fresco.suite.tinytables.util.Encoding;
 import dk.alexandra.fresco.suite.tinytables.util.ot.OTSender;
 import dk.alexandra.fresco.suite.tinytables.util.ot.datatypes.OTInput;
-import edu.biu.scapi.comm.Channel;
 import edu.biu.scapi.exceptions.FactoriesException;
 import edu.biu.scapi.exceptions.SecurityLevelException;
 import edu.biu.scapi.interactiveMidProtocols.ot.otBatch.OTBatchOnByteArraySInput;
 import edu.biu.scapi.interactiveMidProtocols.ot.otBatch.semiHonest.OTSemiHonestDDHBatchOnByteArraySender;
 import edu.biu.scapi.tools.Factories.KdfFactory;
 
+/**
+ * This OTSender is a wrapper around SCAPI's
+ * {@link OTSemiHonestDDHBatchOnByteArraySender} based on an elliptic curve over
+ * a finite field, namely the K-163 curve.
+ * 
+ * @author Jonas Lindstrøm (jonas.lindstrom@alexandra.dk)
+ *
+ */
 public class JavaOTSender implements OTSender {
 
 	private Network network;
 	private int myId;
 	private SecureRandom random;
 
+	/*
+	 * We keep a singleton of the actual sender.
+	 */
 	private static OTSemiHonestDDHBatchOnByteArraySender sender;
 	private OTSemiHonestDDHBatchOnByteArraySender getInstance(SecureRandom random) {
 		if (sender == null) {
 			try {
 				sender = new OTSemiHonestDDHBatchOnByteArraySender(
-						new edu.biu.scapi.primitives.dlog.bc.BcDlogECF2m(), KdfFactory.getInstance()
+						//new edu.biu.scapi.primitives.dlog.bc.BcDlogECF2m(), 
+						new edu.biu.scapi.primitives.dlog.openSSL.OpenSSLDlogECF2m(), 
+						KdfFactory.getInstance()
 						.getObject("HKDF(HMac(SHA-256))"), random);
-			} catch (SecurityLevelException | IOException | FactoriesException e) {
+			} catch (SecurityLevelException | FactoriesException | IOException e) {
 				e.printStackTrace();
 			}
 		}
+		
 		return sender;
 	}
 	
@@ -56,43 +68,15 @@ public class JavaOTSender implements OTSender {
 			x1.add(x1i);
 		}
 		
-		
 		OTBatchOnByteArraySInput otsInputs = new OTBatchOnByteArraySInput(x0, x1);
 		OTSemiHonestDDHBatchOnByteArraySender sender = getInstance(random);
 		
 		try {
-			sender.transfer(new Channel() {
-
-				@Override
-				public void close() {
-					
-				}
-
-				@Override
-				public boolean isClosed() {
-					return false;
-				}
-
-				@Override
-				public Serializable receive() throws ClassNotFoundException, IOException {
-					return network.receive("0", otherPlayerId());
-				}
-
-				@Override
-				public void send(Serializable otInputs) throws IOException {
-					network.send("0", otherPlayerId(), otInputs);
-				}
-				
-			}, otsInputs);
+			sender.transfer(new NetworkWrapper(network, myId), otsInputs);
 		} catch (ClassNotFoundException | IOException e) {
 			// Do nothing
 		}
 		
 	}
-
-	private int otherPlayerId() {
-		return myId == 1 ? 2 : 1;
-	}
-
 	
 }
